API

Standalone

Regression

Standard constructors

MLJLinearModels.LinearRegression — Function

LinearRegression(; fit_intercept)

Objective function: $|Xθ - y|₂²/2$.

MLJLinearModels.RidgeRegression — Function

RidgeRegression()
RidgeRegression(λ; lambda, fit_intercept, penalize_intercept, scale_penalty_with_samples)

Objective function: $|Xθ - y|₂²/2 + n⋅λ|θ|₂²/2$, where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₂²/2 + λ|θ|₂²/2$.

MLJLinearModels.LassoRegression — Function

LassoRegression()
LassoRegression(λ; lambda, fit_intercept, penalize_intercept, scale_penalty_with_samples)

Objective function: $|Xθ - y|₂²/2 + n⋅λ|θ|₁$, where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₂²/2 + λ|θ|₁$

MLJLinearModels.ElasticNetRegression — Function

ElasticNetRegression()
ElasticNetRegression(λ)
ElasticNetRegression(λ, γ; lambda, gamma, fit_intercept, penalize_intercept, scale_penalty_with_samples)

Objective function: $|Xθ - y|₂²/2 + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$, where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₂²/2 + λ|θ|₂²/2 + γ|θ|₁$

MLJLinearModels.HuberRegression — Function

HuberRegression()
HuberRegression(δ)
HuberRegression(δ, λ)
HuberRegression(δ, λ, γ; delta, lambda, gamma, penalty, fit_intercept, scale_penalty_with_samples, penalize_intercept)

Huber Regression with objective:

$∑ρ(Xθ - y) + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$

Where ρ is the Huber function ρ(r) = r²/2if|r|≤δandρ(r)=δ(|r|-δ/2)otherwise andnis the number of samplessize(X, 1). Withscalepenaltywith_samples = falsethe objective function is∑ρ(Xθ - y) + λ|θ|₂²/2 + γ|θ|₁`.

MLJLinearModels.QuantileRegression — Function

QuantileRegression()
QuantileRegression(δ)
QuantileRegression(δ, λ)
QuantileRegression(δ, λ, γ; delta, lambda, gamma, penalty, fit_intercept, scale_penalty_with_samples, penalize_intercept)

Quantile Regression with objective:

$∑ρ(Xθ - y) + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$

Where ρ is the check function ρ(r) = r(δ - 1(r < 0)) and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $∑ρ(Xθ - y) + λ|θ|₂²/2 + γ|θ|₁$.

MLJLinearModels.LADRegression — Function

LADRegression()
LADRegression(λ)
LADRegression(λ, γ; lambda, gamma, penalty, scale_penalty_with_samples, fit_intercept, penalize_intercept)

Least Absolute Deviation regression with objective:

$|Xθ - y|₁ + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$ where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₁ + λ|θ|₂²/2 + γ|θ|₁$.

This is a specific type of Quantile Regression with δ=0.5 (median).

Generic constructors

MLJLinearModels.GeneralizedLinearRegression — Type

GeneralizedLinearRegression{L<:Loss, P<:Penalty}

Generalized Linear Regression (GLR) model with objective function:

$L(y, Xθ) + P(θ)$

where L is a loss function, P a penalty, y is the vector of observed response, X is the feature matrix and θ the vector of parameters. If scale_penalty_with_samples = true (default) the penalty is automatically scaled with the number of samples.

Special cases include:

OLS regression: L2 loss, no penalty.

Ridge regression: L2 loss, L2 penalty.

Lasso regression: L2 loss, L1 penalty.

Logistic regression: Logit loss, [no,L1,L2] penalty.

MLJLinearModels.RobustRegression — Function

RobustRegression()
RobustRegression(ρ)
RobustRegression(ρ, λ)
RobustRegression(ρ, λ, γ; rho, lambda, gamma, penalty, fit_intercept, scale_penalty_with_samples, penalize_intercept)

Objective function: $∑ρ(Xθ - y) + n⋅λ|θ|₂² + n⋅γ|θ|₁$ where ρ is a given function on the residuals and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $∑ρ(Xθ - y) + λ|θ|₂² + γ|θ|₁$.

Classification

MLJLinearModels.LogisticRegression — Function

LogisticRegression()
LogisticRegression(λ)
LogisticRegression(λ, γ; lambda, gamma, penalty, fit_intercept, penalize_intercept, scale_penalty_with_samples, multi_class, nclasses)

Objective function: $L(y, Xθ) + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$ where L is either the logistic loss in the binary case or the multinomial loss otherwise and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $L(y, Xθ) + λ|θ|₂²/2 + γ|θ|₁$.

MLJLinearModels.MultinomialRegression — Function

MultinomialRegression(a; kwa...)

Objective function: $L(y, Xθ) + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$ where L is the multinomial loss and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $L(y, Xθ) + λ|θ|₂²/2 + γ|θ|₁$.

MLJ Interface

Regressors

MLJLinearModels.LinearRegressor — Type

Standard linear regression model with objective function

$|Xθ - y|₂²/2$

Parameters

fit_intercept::Bool

whether to fit the intercept or not. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use (if nothing the default is used). The solver is Cholesky by default but can be Conjugate-Gradient as well. See ?Analytical for more information. Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(LinearRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.RidgeRegressor — Type

Ridge regression model with objective function

$|Xθ - y|₂²/2 + n⋅λ|θ|₂²/2$

where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₂²/2 + λ|θ|₂²/2$.

Parameters

lambda::Real

strength of the L2 regularisation. Default: 1.0

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use (if nothing the default is used). The solver is Cholesky by default but can be Conjugate-Gradient as well. See ?Analytical for more information. Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(RidgeRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.LassoRegressor — Type

Lasso regression model with objective function

$|Xθ - y|₂²/2 + n⋅λ|θ|₁$

where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₂²/2 + λ|θ|₁$

Parameters

lambda::Real

strength of the L1 regularisation. Default: 1.0

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use (if nothing the default is used). Either FISTA or ISTA can be used (proximal methods, with/without acceleration). Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(LassoRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.ElasticNetRegressor — Type

Elastic net regression model with objective function

$|Xθ - y|₂²/2 + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$

where $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $|Xθ - y|₂²/2 + λ|θ|₂²/2 + γ|θ|₁$

Parameters

lambda::Real

strength of the L2 regularisation. Default: 1.0

gamma::Real

strength of the L1 regularisation. Default: 0.0

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use (if nothing the default is used). Either FISTA or ISTA can be used (proximal methods, with/without acceleration). Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(ElasticNetRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.HuberRegressor — Type

Huber Regression is the same as RobustRegressor but with the robust loss set to HuberRho.

Parameters

delta::Real

parametrises the HuberRho function (radius of the ball within which the loss is a quadratic loss) Default: 0.5

lambda::Real

strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en. Default: 1.0

gamma::Real

strength of the L1 regulariser if penalty is :en. Default: 0.0

penalty::Union{String, Symbol}

the penalty to use, either :l2, :l1, :en (elastic net) or :none. Default: :l2

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use, default if nothing. Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(HuberRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.QuantileRegressor — Type

Quantile Regression is the same as RobustRegressor but with the robust loss set to QuantileRho.

Parameters

delta::Real

parametrises the QuantileRho function (indicating the quantile to use with default 0.5 for the median regression) Default: 0.5

lambda::Real

strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en. Default: 1.0

gamma::Real

strength of the L1 regulariser if penalty is :en. Default: 0.0

penalty::Union{String, Symbol}

the penalty to use, either :l2, :l1, :en (elastic net) or :none. Default: :l2

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use, default if nothing. Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(QuantileRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.LADRegressor — Type

Least Absolute Deviation regression with with objective function

$∑ρ(Xθ - y) + n⋅λ|θ|₂² + n⋅γ|θ|₁$

where $ρ$ is the absolute loss and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $∑ρ(Xθ - y) + λ|θ|₂² + γ|θ|₁$

See also RobustRegressor.

Parameters

lambda::Real

strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en. Default: 1.0

gamma::Real

strength of the L1 regulariser if penalty is :en. Default: 0.0

penalty::Union{String, Symbol}

the penalty to use, either :l2, :l1, :en (elastic net) or :none. Default: :l2

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use, default if nothing. Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(LADRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.RobustRegressor — Type

Robust regression model with objective function

$∑ρ(Xθ - y) + n⋅λ|θ|₂² + n⋅γ|θ|₁$

where $ρ$ is a robust loss function (e.g. the Huber function) and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $∑ρ(Xθ - y) + λ|θ|₂² + γ|θ|₁$.

Parameters

rho::MLJLinearModels.RobustRho

the type of robust loss to use (see HuberRho, TalwarRho, ...) Default: HuberRho(0.1)

lambda::Real

strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en. Default: 1.0

gamma::Real

strength of the L1 regulariser if penalty is :en. Default: 0.0

penalty::Union{String, Symbol}

the penalty to use, either :l2, :l1, :en (elastic net) or :none. Default: :l2

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use, default if nothing. Default: nothing

Example

using MLJ
X, y = make_regression()
mach = fit!(machine(RobustRegressor(), X, y))
predict(mach, X)
fitted_params(mach)

Classifiers

MLJLinearModels.LogisticClassifier — Type

Logistic Classifier (typically called "Logistic Regression"). This model is a standard classifier for both binary and multiclass classification. In the binary case it corresponds to the LogisticLoss, in the multiclass to the Multinomial (softmax) loss. An elastic net penalty can be applied with overall objective function

$L(y, Xθ) + n⋅λ|θ|₂²/2 + n⋅γ|θ|₁$

where $L$ is either the logistic or multinomial loss and $λ$ and $γ$ indicate the strength of the L2 (resp. L1) regularisation components and $n$ is the number of samples size(X, 1). With scale_penalty_with_samples = false the objective function is $L(y, Xθ) + λ|θ|₂²/2 + γ|θ|₁$

Parameters

lambda::Real

strength of the regulariser if penalty is :l2 or :l1 and strength of the L2 regulariser if penalty is :en. Default: eps()

gamma::Real

strength of the L1 regulariser if penalty is :en. Default: 0.0

penalty::Union{String, Symbol}

the penalty to use, either :l2, :l1, :en (elastic net) or :none. Default: :l2

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use, default if nothing. Default: nothing

Example

using MLJ
X, y = make_blobs(centers = 2)
mach = fit!(machine(LogisticClassifier(), X, y))
predict(mach, X)
fitted_params(mach)

MLJLinearModels.MultinomialClassifier — Type

See LogisticClassifier, it's the same except that multiple classes are assumed by default. The other parameters are the same.

Parameters

lambda::Real

strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en. Default: eps()

gamma::Real

strength of the L1 regulariser if penalty is :en. Default: 0.0

penalty::Union{String, Symbol}

the penalty to use, either :l2, :l1, :en (elastic net) or :none. Default: :l2

fit_intercept::Bool

whether to fit the intercept or not. Default: true

penalize_intercept::Bool

whether to penalize the intercept. Default: false

scale_penalty_with_samples::Bool

whether to scale the penalty with the number of samples. Default: true

solver::Union{Nothing, MLJLinearModels.Solver}

type of solver to use, default if nothing. Default: nothing

Example

using MLJ
X, y = make_blobs(centers = 3)
mach = fit!(machine(LogisticClassifier(), X, y))
predict(mach, X)
fitted_params(mach)