API

Standalone

Regression

Standard constructors

MLJLinearModels.LinearRegression — Function

LinearRegression(; fit_intercept)

Objective function: $|Xθ - y|₂²/2$.

MLJLinearModels.RidgeRegression — Function

RidgeRegression()
RidgeRegression(λ; lambda, fit_intercept, penalize_intercept)

Objective function: $|Xθ - y|₂²/2 + λ|θ|₂²/2$.

MLJLinearModels.LassoRegression — Function

LassoRegression()
LassoRegression(λ; lambda, fit_intercept, penalize_intercept)

Objective function: $|Xθ - y|₂²/2 + λ|θ|₁$.

MLJLinearModels.ElasticNetRegression — Function

ElasticNetRegression()
ElasticNetRegression(λ)
ElasticNetRegression(λ, γ; lambda, gamma, fit_intercept, penalize_intercept)

Objective function: $|Xθ - y|₂²/2 + λ|θ|₂²/2 + γ|θ|₁$.

MLJLinearModels.HuberRegression — Function

HuberRegression()
HuberRegression(δ)
HuberRegression(δ, λ)
HuberRegression(δ, λ, γ; delta, lambda, gamma, penalty, fit_intercept, penalize_intercept)

Huber Regression with objective:

$∑ρ(Xθ - y) + λ|θ|₂²/2 + γ|θ|₁$

Where ρ is the Huber function ρ(r) = r²/2if|r|≤δandρ(r)=δ(|r|-δ/2)` otherwise.

MLJLinearModels.QuantileRegression — Function

QuantileRegression()
QuantileRegression(δ)
QuantileRegression(δ, λ)
QuantileRegression(δ, λ, γ; delta, lambda, gamma, penalty, fit_intercept, penalize_intercept)

Quantile Regression with objective:

$∑ρ(Xθ - y) + λ|θ|₂²/2 + γ|θ|₁$

Where ρ is the check function ρ(r) = r(δ - 1(r < 0)).

MLJLinearModels.LADRegression — Function

LADRegression()
LADRegression(λ)
LADRegression(λ, γ; lambda, gamma, penalty, fit_intercept, penalize_intercept)

Least Absolute Deviation regression with objective:

$|Xθ - y|₁ + λ|θ|₂²/2 + γ|θ|₁$

This is a specific type of Quantile Regression with δ=0.5 (median).

Generic constructors

MLJLinearModels.GeneralizedLinearRegression — Type

GeneralizedLinearRegression{L<:Loss, P<:Penalty}

Generalized Linear Regression (GLR) model with objective function:

$L(y, Xθ) + P(θ)$

where L is a loss function, P a penalty, y is the vector of observed response, X is the feature matrix and θ the vector of parameters.

Special cases include:

OLS regression: L2 loss, no penalty.

Ridge regression: L2 loss, L2 penalty.

Lasso regression: L2 loss, L1 penalty.

Logistic regression: Logit loss, [no,L1,L2] penalty.

MLJLinearModels.RobustRegression — Function

RobustRegression()
RobustRegression(ρ)
RobustRegression(ρ, λ)
RobustRegression(ρ, λ, γ; rho, lambda, gamma, penalty, fit_intercept, penalize_intercept)

Objective function: $∑ρ(Xθ - y) + λ|θ|₂² + γ|θ|₁$ where ρ is a given function on the residuals.

Classification

MLJLinearModels.LogisticRegression — Function

LogisticRegression()
LogisticRegression(λ)
LogisticRegression(λ, γ; lambda, gamma, penalty, multi_class, fit_intercept, penalize_intercept)

Objective function: $L(y, Xθ) + λ|θ|₂²/2 + γ|θ|₁$ where L is either the logistic loss in the binary case or the multinomial loss otherwise.

MLJLinearModels.MultinomialRegression — Function

MultinomialRegression(a; kwa...)

Objective function: $L(y, Xθ) + λ|θ|₂²/2 + γ|θ|₁$ where L is the multinomial loss.

MLJ Interface

Regressors

MLJLinearModels.LinearRegressor — Type

Standard linear regression model.

Parameters

fit_intercept (Bool): whether to fit the intercept or not.

solver: type of solver to use (if nothing the default is used). The solver is Cholesky by default but can be Conjugate-Gradient as well. See ?Analytical for more information.

MLJLinearModels.RidgeRegressor — Type

Ridge regression model with objective function

$|Xθ - y|₂²/2 + λ|θ|₂²/2$

Parameters

lambda (Real): strength of the L2 regularisation.

fit_intercept (Bool): whether to fit the intercept or not.

penalize_intercept (Bool): whether to penalize the intercept.

solver: type of solver to use (if nothing the default is used). The solver is Cholesky by default but can be Conjugate-Gradient as well. See ?Analytical for more information.

MLJLinearModels.LassoRegressor — Type

Lasso regression model with objective function

$|Xθ - y|₂²/2 + λ|θ|₁$

Parameters

lambda (Real): strength of the L1 regularisation.

fit_intercept (Bool): whether to fit the intercept or not.

penalize_intercept (Bool): whether to penalize the intercept.

solver: type of solver to use (if nothing the default is used). Either FISTA or ISTA can be used (proximal methods, with/without acceleration).

MLJLinearModels.ElasticNetRegressor — Type

Elastic net regression model with objective function

$|Xθ - y|₂²/2 + λ|θ|₂²/2 + γ|θ|₁$

Parameters

lambda (Real): strength of the L2 regularisation.

gamma (Real): strength of the L1 regularisation.

fit_intercept (Bool): whether to fit the intercept or not.

penalize_intercept (Bool): whether to penalize the intercept.

solver: type of solver to use (if nothing the default is used). Either FISTA or ISTA can be used (proximal methods, with/without acceleration).

MLJLinearModels.HuberRegressor — Type

Huber Regression, see RobustRegressor, it's the same but with the robust loss set to HuberRho. The parameters are the same apart from delta which parametrises the HuberRho function (radius of the ball within which the loss is a quadratic loss).

MLJLinearModels.QuantileRegressor — Type

Quantile Regression, see RobustRegressor, it's the same but with the robust loss set to QuantileRho. The parameters are the same apart from delta which parametrises the QuantileRho function (indicating the quantile to use with default 0.5 for the median regression).

MLJLinearModels.LADRegressor — Type

Least Absolute Deviation regression with with objective function

$∑ρ(Xθ - y) + λ|θ|₂² + γ|θ|₁$

where ρ is the absolute loss.

See also RobustRegressor.

MLJLinearModels.RobustRegressor — Type

Robust regression model with objective function

$∑ρ(Xθ - y) + λ|θ|₂² + γ|θ|₁$

where ρ is a robust loss function (e.g. the Huber function).

Parameters

rho (RobustRho): the type of robust loss to use (see HuberRho, TalwarRho, ...)

penalty (Symbol or String): the penalty to use, either :l2, :l1, :en (elastic net) or :none. (Default: :l2)

lambda (Real): strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en.

gamma (Real): strength of the L1 regulariser if penalty is :en.

fit_intercept (Bool): whether to fit an intercept (Default: true)

penalize_intercept (Bool): whether to penalize intercept (Default: false)

solver (Solver): type of solver to use, default if nothing.

Classifiers

MLJLinearModels.LogisticClassifier — Type

Logistic Classifier (typically called "Logistic Regression"). This model is a standard classifier for both binary and multiclass classification. In the binary case it corresponds to the LogisticLoss, in the multiclass to the Multinomial (softmax) loss. An elastic net penalty can be applied with overall objective function

$L(y, Xθ) + λ|θ|₂²/2 + γ|θ|₁$

Where L is either the logistic or multinomial loss and λ and γ indicate the strength of the L2 (resp. L1) regularisation components.

Parameters

penalty (Symbol or String): the penalty to use, either :l2, :l1, :en (elastic net) or :none. (Default: :l2)

lambda (Real): strength of the regulariser if penalty is :l2 or :l1. Strength of the L2 regulariser if penalty is :en.

gamma (Real): strength of the L1 regulariser if penalty is :en.

fit_intercept (Bool): whether to fit an intercept (Default: true)

penalize_intercept (Bool): whether to penalize intercept (Default: false)

solver (Solver): type of solver to use, default if nothing.

multi_class (Bool): whether it's a binary or multi class classification problem. This is usually set automatically.

MLJLinearModels.MultinomialClassifier — Type

See LogisticClassifier, it's the same except that multi_class is set to true by default. The other parameters are the same.